Address review comments
 - Created a separate python script for image classifier model definition and
   imported it in the flow inside `train` and `evalauate` step
 - Changed Flow file name and class name to include the word `flow`
