+++ examples/image_classifier/image_classifier_flow.py
@@ -23,7 +23,7 @@ GCS_BUCKET_NAME = "your-gcs-bucket-here"
 MODEL_STORAGE_PATH = "image_classifier/trained-model-bytes.pth"
 
 
-class ImageClassifier(FlowSpec):
+class ImageClassifierFlow(FlowSpec):
     # This is an example of a parameter. You can toggle this when you call the flow
     # with python template_flow.py run --offline False
     offline_wandb = Parameter(
@@ -77,8 +77,8 @@ class ImageClassifier(FlowSpec):
     def train(self):
         import torch
         import torch.nn as nn
-        import torch.nn.functional as F
         import torch.optim as optim
+        from image_classifier_model import ImageClassifierModel
         from io import BytesIO
         import wandb
         import os
@@ -101,31 +101,13 @@ class ImageClassifier(FlowSpec):
 
         print(f"Training on: {device}")
 
-        # Define a Convolutional Neural Network
-        class Net(nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.conv1 = nn.Conv2d(3, 6, 5)
-                self.pool = nn.MaxPool2d(2, 2)
-                self.conv2 = nn.Conv2d(6, 16, 5)
-                self.fc1 = nn.Linear(16 * 5 * 5, 120)
-                self.fc2 = nn.Linear(120, 84)
-                self.fc3 = nn.Linear(84, 10)
-
-            def forward(self, x):
-                x = self.pool(F.relu(self.conv1(x)))
-                x = self.pool(F.relu(self.conv2(x)))
-                x = torch.flatten(x, 1)  # flatten all dimensions except batch
-                x = F.relu(self.fc1(x))
-                x = F.relu(self.fc2(x))
-                x = self.fc3(x)
-                return x
-
-        net = Net().to(device)
+        image_classifier_model = ImageClassifierModel().to(device)
 
         # Define a Loss function and optimizer
         criterion = nn.CrossEntropyLoss()
-        optimizer = optim.SGD(net.parameters(), lr=0.001, momentum=0.9)
+        optimizer = optim.SGD(
+            image_classifier_model.parameters(), lr=0.001, momentum=0.9
+        )
 
         # Load train data
         batch_size = 4
@@ -145,7 +127,7 @@ class ImageClassifier(FlowSpec):
                 optimizer.zero_grad()
 
                 # forward + backward + optimize
-                outputs = net(inputs)
+                outputs = image_classifier_model(inputs)
                 loss = criterion(outputs, labels)
                 loss.backward()
                 optimizer.step()
@@ -160,7 +142,7 @@ class ImageClassifier(FlowSpec):
 
         print("Finished Training")
         buffer = BytesIO()
-        torch.save(net.state_dict(), buffer)
+        torch.save(image_classifier_model.state_dict(), buffer)
         self.model_state_dict_bytes = buffer.getvalue()
         self.next(self.evaluate)
 
@@ -176,36 +158,17 @@ class ImageClassifier(FlowSpec):
     @step
     def evaluate(self):
         import torch
-        import torch.nn as nn
-        import torch.nn.functional as F
+        from image_classifier_model import ImageClassifierModel
         from io import BytesIO
 
         device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
         print(f"Evaluating on: {device}")
 
-        # Define a Convolutional Neural Network
-        class Net(nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.conv1 = nn.Conv2d(3, 6, 5)
-                self.pool = nn.MaxPool2d(2, 2)
-                self.conv2 = nn.Conv2d(6, 16, 5)
-                self.fc1 = nn.Linear(16 * 5 * 5, 120)
-                self.fc2 = nn.Linear(120, 84)
-                self.fc3 = nn.Linear(84, 10)
-
-            def forward(self, x):
-                x = self.pool(F.relu(self.conv1(x)))
-                x = self.pool(F.relu(self.conv2(x)))
-                x = torch.flatten(x, 1)  # flatten all dimensions except batch
-                x = F.relu(self.fc1(x))
-                x = F.relu(self.fc2(x))
-                x = self.fc3(x)
-                return x
-
-        net = Net().to(device)
+        image_classifier_model = ImageClassifierModel().to(device)
         buffer = BytesIO(self.model_state_dict_bytes)
-        net.load_state_dict(torch.load(buffer, map_location=device, weights_only=True))
+        image_classifier_model.load_state_dict(
+            torch.load(buffer, map_location=device, weights_only=True)
+        )
 
         correct = 0
         total = 0
@@ -220,7 +183,7 @@ class ImageClassifier(FlowSpec):
             for data in testloader:
                 images, labels = data[0].to(device), data[1].to(device)
                 # calculate outputs by running images through the network
-                outputs = net(images)
+                outputs = image_classifier_model(images)
                 # the class with the highest energy is what we choose as prediction
                 _, predicted = torch.max(outputs.data, 1)
                 total += labels.size(0)
@@ -260,4 +223,4 @@ class ImageClassifier(FlowSpec):
 
 
 if __name__ == "__main__":
-    ImageClassifier()
+    ImageClassifierFlow()
